windows: always enable concurrent image pull
Move the logic for determining if concurrent image pull should be
enabled to a linux specific file. This should always be enabled in
windows on account of the bug being fixed in releases prior to windows
support.
